requirements.txt
boto3
requests_aws4auth
elasticsearch-curator
elasticsearch